requirements.txt
python-dotenv
python-telegram-bot~=13.13
redis
